Made pickers for lat/lng smaller.
